requirements.txt
langchain
python-dotenv
openai
tiktoken
streamlit
streamlit-ace
backoff
promptlayer
beautifulsoup4
requests
requests-html
tiktoken